+++ changelog.md
@@ -1,2 +1,2 @@
 ### Tesseract 1.0.38
-- Initial release of Tesseract for Minecraft 1.21.9 & 1.21.10
+- Initial release of Tesseract for Minecraft 1.21.11